requirements.txt
# Wealth Seeker v0.01 - Dependencies

# Core Trading & Data
requests>=2.31.0
yfinance>=0.2.35
pandas>=2.1.0
numpy>=1.24.0

# AI/ML
google-generativeai>=0.3.0

# Web & UI (ApexCharts loaded via CDN in HTML)
jinja2>=3.1.2

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
python-dotenv>=1.0.0

# Notifications
python-telegram-bot>=20.7

# Flask Application
Flask>=3.0.0
Flask-CORS>=4.0.0
gunicorn>=21.0.0
psutil>=5.9.0

# Athena & Alternative Data (Added Feb 2026)
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
finnhub-python>=2.4.19
pytrends>=4.9.2
playwright>=1.41.0
groq>=0.4.0
beautifulsoup4>=4.12.3
krakenex>=2.2.0
pycryptodome>=3.18.0
